Housing foundation - hand stitched
Creating foundation for Feature/Housing hand stitched with agatho's work
Fix31

+++ src/server/game/DataStores/DB2Structure.h
@@ -5037,7 +5037,7 @@ struct HouseExteriorWmoDataEntry
 struct HouseLevelDataEntry  
 {  
     uint32 ID;  
-    int32 Level;  
+    int8 Level;  
     int32 QuestID;  
     int32 Field_12_0_7_67808_003;  
 };
